Re-organized the tests
