+++ docs/unicode.rst
@@ -1,0 +1,6 @@
+Unicode functions
+=================
+
+.. doxygenenum:: skyr::unicode::idna_status
+
+.. doxygenenum:: skyr::unicode::unicode_errc
