+++ docs/url_error_codes.rst
